+++ packages/@baldeweg/ui/src/components/BSubheader/BSubheader.stories.js
@@ -12,7 +12,7 @@ export default {
       },
     },
   },
-  tags: ['experimental'],
+  tags: ['deprecated'],
 }
 
 const Template = (args) => ({