+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Improved color contrast
 - Simplified `BDivider` component by removing size prop and related styles
 - `BSubheader` component is deprecated
+- `BFlex` component is deprecated
 
 ## v0.15.0